refactor: StatusCode 패키지 이동
